+++ pkg/cluster/controller/groups/members/zz_controller_test.go
@@ -39,9 +39,9 @@ import (
 var (
 	unexpecedItem resource.Managed
 	errBoom       = errors.New("boom")
-	ID            = 0
+	ID            = int64(0)
 	username      = "username"
-	userID        = 123
+	userID        = int64(123)
 	name          = "name"
 	state         = "state"
 	avatarURL     = "http://avatarURL"
@@ -50,7 +50,7 @@ var (
 	now           = time.Now()
 	expiresAt     = gitlab.ISOTime(now.AddDate(0, 0, 7*3))
 	expiresAtNew  = gitlab.ISOTime(now.AddDate(0, 0, 7*4))
-	groupID       = 1234
+	groupID       = int64(1234)
 )
 
 type args struct {
@@ -165,7 +165,7 @@ func TestObserve(t *testing.T) {
 		"FailedGetRequest": {
 			args: args{
 				groupMember: &fake.MockClient{
-					MockGetMember: func(gid interface{}, user int, options ...gitlab.RequestOptionFunc) (*gitlab.GroupMember, *gitlab.Response, error) {
+					MockGetMember: func(gid interface{}, user int64, options ...gitlab.RequestOptionFunc) (*gitlab.GroupMember, *gitlab.Response, error) {
 						return nil, &gitlab.Response{Response: &http.Response{StatusCode: 400}}, errBoom
 					},
 				},
@@ -184,7 +184,7 @@ func TestObserve(t *testing.T) {
 		"ErrGet404": {
 			args: args{
 				groupMember: &fake.MockClient{
-					MockGetMember: func(gid interface{}, user int, options ...gitlab.RequestOptionFunc) (*gitlab.GroupMember, *gitlab.Response, error) {
+					MockGetMember: func(gid interface{}, user int64, options ...gitlab.RequestOptionFunc) (*gitlab.GroupMember, *gitlab.Response, error) {
 						return nil, &gitlab.Response{Response: &http.Response{StatusCode: 404}}, errBoom
 					},
 				},
@@ -203,7 +203,7 @@ func TestObserve(t *testing.T) {
 		"SuccessfulAvailable": {
 			args: args{
 				groupMember: &fake.MockClient{
-					MockGetMember: func(gid interface{}, user int, options ...gitlab.RequestOptionFunc) (*gitlab.GroupMember, *gitlab.Response, error) {
+					MockGetMember: func(gid interface{}, user int64, options ...gitlab.RequestOptionFunc) (*gitlab.GroupMember, *gitlab.Response, error) {
 						return &gitlab.GroupMember{}, &gitlab.Response{}, nil
 					},
 				},
@@ -228,7 +228,7 @@ func TestObserve(t *testing.T) {
 		"IsGroupUpToDateAccessLevel": {
 			args: args{
 				groupMember: &fake.MockClient{
-					MockGetMember: func(gid interface{}, user int, options ...gitlab.RequestOptionFunc) (*gitlab.GroupMember, *gitlab.Response, error) {
+					MockGetMember: func(gid interface{}, user int64, options ...gitlab.RequestOptionFunc) (*gitlab.GroupMember, *gitlab.Response, error) {
 						return &gitlab.GroupMember{
 							AccessLevel: accessLevel,
 						}, &gitlab.Response{}, nil
@@ -257,7 +257,7 @@ func TestObserve(t *testing.T) {
 		"IsGroupUpToDateExpiresAt": {
 			args: args{
 				groupMember: &fake.MockClient{
-					MockGetMember: func(gid interface{}, user int, options ...gitlab.RequestOptionFunc) (*gitlab.GroupMember, *gitlab.Response, error) {
+					MockGetMember: func(gid interface{}, user int64, options ...gitlab.RequestOptionFunc) (*gitlab.GroupMember, *gitlab.Response, error) {
 						return &gitlab.GroupMember{ExpiresAt: &expiresAt}, &gitlab.Response{}, nil
 					},
 				},
@@ -284,7 +284,7 @@ func TestObserve(t *testing.T) {
 		"NoUserIDandNoUserName": {
 			args: args{
 				groupMember: &fake.MockClient{
-					MockGetMember: func(gid interface{}, user int, options ...gitlab.RequestOptionFunc) (*gitlab.GroupMember, *gitlab.Response, error) {
+					MockGetMember: func(gid interface{}, user int64, options ...gitlab.RequestOptionFunc) (*gitlab.GroupMember, *gitlab.Response, error) {
 						return nil, nil, errBoom
 					},
 				},
@@ -308,7 +308,7 @@ func TestObserve(t *testing.T) {
 		"NoUserIDSuccess": {
 			args: args{
 				groupMember: &fake.MockClient{
-					MockGetMember: func(gid interface{}, user int, options ...gitlab.RequestOptionFunc) (*gitlab.GroupMember, *gitlab.Response, error) {
+					MockGetMember: func(gid interface{}, user int64, options ...gitlab.RequestOptionFunc) (*gitlab.GroupMember, *gitlab.Response, error) {
 						return &gitlab.GroupMember{}, &gitlab.Response{}, nil
 					},
 				},
@@ -504,7 +504,7 @@ func TestUpdate(t *testing.T) {
 		"SuccessfulUpdate": {
 			args: args{
 				groupMember: &fake.MockClient{
-					MockEditMember: func(gid interface{}, user int, opt *gitlab.EditGroupMemberOptions, options ...gitlab.RequestOptionFunc) (*gitlab.GroupMember, *gitlab.Response, error) {
+					MockEditMember: func(gid interface{}, user int64, opt *gitlab.EditGroupMemberOptions, options ...gitlab.RequestOptionFunc) (*gitlab.GroupMember, *gitlab.Response, error) {
 						return &gitlab.GroupMember{
 							ID:          ID,
 							Username:    username,
@@ -533,7 +533,7 @@ func TestUpdate(t *testing.T) {
 		"FailedUpdate": {
 			args: args{
 				groupMember: &fake.MockClient{
-					MockEditMember: func(gid interface{}, user int, opt *gitlab.EditGroupMemberOptions, options ...gitlab.RequestOptionFunc) (*gitlab.GroupMember, *gitlab.Response, error) {
+					MockEditMember: func(gid interface{}, user int64, opt *gitlab.EditGroupMemberOptions, options ...gitlab.RequestOptionFunc) (*gitlab.GroupMember, *gitlab.Response, error) {
 						return &gitlab.GroupMember{}, &gitlab.Response{}, errBoom
 					},
 				},
@@ -589,7 +589,7 @@ func TestDelete(t *testing.T) {
 		"SuccessfulDeletion": {
 			args: args{
 				groupMember: &fake.MockClient{
-					MockRemoveMember: func(gid interface{}, user int, options ...gitlab.RequestOptionFunc) (*gitlab.Response, error) {
+					MockRemoveMember: func(gid interface{}, user int64, options ...gitlab.RequestOptionFunc) (*gitlab.Response, error) {
 						return &gitlab.Response{}, nil
 					},
 				},
@@ -607,7 +607,7 @@ func TestDelete(t *testing.T) {
 		"FailedDeletion": {
 			args: args{
 				groupMember: &fake.MockClient{
-					MockRemoveMember: func(gid interface{}, user int, options ...gitlab.RequestOptionFunc) (*gitlab.Response, error) {
+					MockRemoveMember: func(gid interface{}, user int64, options ...gitlab.RequestOptionFunc) (*gitlab.Response, error) {
 						return &gitlab.Response{}, errBoom
 					},
 				},
