+++ pkg/cluster/controller/instance/runners/zz_controller_test.go
@@ -48,7 +48,7 @@ var (
 	unexpecedItem resource.Managed
 
 	errBoom           = errors.New("boom")
-	runnerID          = 1
+	runnerID          = int64(1)
 	extName           = "1"
 	extNameAnnotation = map[string]string{meta.AnnotationKeyExternalName: extName}
 )
@@ -356,7 +356,7 @@ func TestDelete(t *testing.T) {
 		"FailedDeletion": {
 			args: args{
 				runnerClient: &runnersfake.MockClient{
-					MockDeleteRegisteredRunnerByID: func(rid int, options ...gitlab.RequestOptionFunc) (*gitlab.Response, error) {
+					MockDeleteRegisteredRunnerByID: func(rid int64, options ...gitlab.RequestOptionFunc) (*gitlab.Response, error) {
 						return &gitlab.Response{}, errBoom
 					},
 				},
@@ -376,7 +376,7 @@ func TestDelete(t *testing.T) {
 		"SuccessfulDeletion": {
 			args: args{
 				runnerClient: &runnersfake.MockClient{
-					MockDeleteRegisteredRunnerByID: func(rid int, options ...gitlab.RequestOptionFunc) (*gitlab.Response, error) {
+					MockDeleteRegisteredRunnerByID: func(rid int64, options ...gitlab.RequestOptionFunc) (*gitlab.Response, error) {
 						return &gitlab.Response{}, nil
 					},
 				},
