+++ pkg/namespaced/clients/projects/hook_test.go
@@ -48,7 +48,7 @@ var (
 )
 
 func TestGenerateHookObservation(t *testing.T) {
-	id := 0
+	id := int64(0)
 	createdAt := time.Now()
 
 	type args struct {
